Relocate files to match newly defined structure
