--- examples/StockDonor/.gitignore
@@ -1,4 +1,0 @@
-node_modules
-.DS_Store
-dist
-yarn-error.log
