--- examples/StockDonor/package.json
@@ -1,19 +1,0 @@
-{
-  "//": "This is also plugin code, and must snake case (since it's used as federation module name",
-  "name": "plugin_example_stock_donor",
-  "//": "Must be set for plugin generation CLI to consider this file as plugin",
-  "isOpenMsupplyPlugin": true,
-  "//": "To do, about version",
-  "version": "1.0.0",
-  "main": "plugin.js",
-  "scripts": {
-    "build-plugin": "NODE_ENV=production webpack"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "webpack": "^5.76.2",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.12.0",
-    "ts-loader": "^9.4.4"
-  }
-}
