--- examples/StockDonor/src/plugin.tsx
@@ -1,10 +1,0 @@
-import { Plugins } from '@openmsupply-client/common';
-import StockDonorEdit from './StockDonorEdit';
-import { StateLoader, StockDonorColumn } from './StockDonorColumn';
-
-const StockDonor: Plugins = {
-  stockEditForm: [StockDonorEdit],
-  stockColumn: { StateLoader: [StateLoader], columns: [StockDonorColumn] },
-};
-
-export default StockDonor;
