--- examples/StockDonor/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "typeRoots": [
-      "./typings",
-      "./node_modules/@types",
-      "../../../node_modules/@types",
-      "../../../typings"
-    ]
-  }
-}
