+++ frontend/latest/src/Dashboard/ReplenishmentWidget.tsx
