+++ frontend/latest/src/Dashboard/SyncStatusWidget.tsx
