+++ frontend/latest/src/Dashboard/api/hooks/statistics/index.ts
