+++ frontend/latest/src/Dashboard/api/hooks/statistics/useRequisitionCounts.ts
