+++ frontend/latest/src/Dashboard/plugin.tsx
