+++ frontend/latest/src/StockDonor/api/hooks/document/useUpdatePluginData.ts
