+++ frontend/latest/src/StockDonor/api/operations.graphql
