+++ frontend/latest/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "//": "Expecting for every plugin to be developed from 'common/packages/plugins/{plugin repo}/frontend/{version}'",
+  "extends": "../../../../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "typeRoots": [
+      "./typings",
+      "./node_modules/@types",
+      "../../../../../node_modules/@types",
+      "../../../../../typings"
+    ]
+  }
+}
