+++ frontend/latest/webpack.config.js
@@ -1,0 +1,13 @@
+const { merge } = require('webpack-merge');
+// Expecting for every plugin to be developed from 'common/packages/plugins/{plugin repo}/frontend/{version}'
+const generateBaseConfig = require('../../../plugin-webpack.config.js');
+const path = require('path');
+
+const pluginName = require('./package.json').name;
+const distDir = path.resolve(__dirname, 'dist');
+const extendConfig = {};
+
+module.exports = merge(
+  generateBaseConfig({ pluginName, distDir }),
+  extendConfig
+);
